+++ example/spec/serial/ingestPdrWithNodeNameFtpSpec.js
@@ -76,7 +76,6 @@ describe('Ingesting from PDR', () => {
       config = await loadConfig();
       functionName = `${config.stackName}-populateTestLambda`;
       process.env.ExecutionsTable = `${config.stackName}-ExecutionsTable`;
-      process.env.PdrsTable = `${config.stackName}-PdrsTable`;
 
       const testId = createTimestampedTestId(config.stackName, 'IngestFromPdrWithNodeName');
       testSuffix = createTestSuffix(testId);