+++ packages/api/models/index.js
@@ -3,9 +3,7 @@
 const AccessToken = require('./access-tokens');
 const AsyncOperation = require('./async-operation');
 const Manager = require('./base');
-const Collection = require('./collections');
 const Granule = require('./granules');
-const Pdr = require('./pdrs');
 const Provider = require('./providers');
 const ReconciliationReport = require('./reconciliation-reports');
 const Rule = require('./rules');
@@ -14,9 +12,7 @@ const Execution = require('./executions');
 module.exports = {
   AccessToken,
   AsyncOperation,
-  Collection,
   Granule,
-  Pdr,
   Provider,
   ReconciliationReport,
   Rule,