+++ packages/api/tests/endpoints/test-reconciliation-reports.js
@@ -37,7 +37,6 @@ process.env.invoke = 'granule-reconciliation-reports';
 process.env.stackName = 'test-stack';
 process.env.system_bucket = 'testsystembucket';
 process.env.AccessTokensTable = randomId('accessTokensTable');
-process.env.AsyncOperationsTable = randomId('asyncOperationsTable');
 process.env.ReconciliationReportsTable = randomId('recReportsTable');
 process.env.TOKEN_SECRET = randomId('tokenSecret');
 process.env.stackName = randomId('stackname');
@@ -62,7 +61,6 @@ const esIndex = randomId('esindex');
 
 let jwtAuthToken;
 let accessTokenModel;
-let asyncOperationsModel;
 let reconciliationReportModel;
 let fakeReportRecords = [];
 
@@ -84,12 +82,6 @@ test.before(async () => {
 
   reconciliationReportModel = new models.ReconciliationReport();
   await reconciliationReportModel.createTable();
-
-  asyncOperationsModel = new models.AsyncOperation({
-    stackName: process.env.stackName,
-    systemBucket: process.env.system_bucket,
-  });
-  await asyncOperationsModel.createTable();
 
   await awsServices.s3().createBucket({
     Bucket: process.env.system_bucket,
@@ -138,7 +130,6 @@ test.before(async () => {
 test.after.always(async () => {
   await accessTokenModel.deleteTable();
   await reconciliationReportModel.deleteTable();
-  await asyncOperationsModel.deleteTable();
   await esClient.indices.delete({
     index: esIndex,
   });